backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# HTTP client
httpx==0.28.1

# GIS
shapely==2.0.6
pyproj==3.7.0

# Data validation
pydantic==2.10.4

# Database (SQLite locally, Postgres for production)
sqlalchemy==2.0.36
aiosqlite==0.20.0
# For production (Railway/Supabase):
# asyncpg==0.30.0
# psycopg2-binary==2.9.10

# LLM
anthropic==0.42.0

# PDF parsing (for CP-7150)
pdfplumber==0.11.4

# Utilities
python-dotenv==1.0.1
orjson==3.10.12

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
